Consider creating a bundler gem

+++ Rakefile
@@ -1,9 +1,15 @@
 # frozen_string_literal: true
 
-require 'fileutils'
+require 'rake/testtask'
+require './file_helper'
 
-LOAD_PATH = File.dirname(File.expand_path(__FILE__))
-BACKUP_DIR = create_backup_dir
+# LOAD_PATH = File.dirname(File.expand_path(__FILE__))
+@fh = FileHelper.new
+BACKUP_DIR = @fh.create_backup_dir
+
+task :test do
+
+end
 
 task :example do
   # p LOAD_PATH
@@ -16,38 +22,21 @@ task :copy_config do
     dot_file = ".#{file}"
     backup_file = "#{dot_file}.orig"
 
-    puts "#{dot_file} does not exist, no backup created" unless file_exists?(dot_file)
+    # Checks that there is a file to create
+    unless file_exists?(dot_file)
+      puts "#{file_path(dot_file)} does not exist, no backup created"
+      next
+    end
 
-    file_exist?(backup_file)
+    if file_exist?(backup_file)
+      puts "#{file_path(backup_file, BACKUP_DIR)} already exists, no backup created"
+      next
+    end
+
     # Copies to the backup dir if a .examplerc exists
     copy_file(dot_file, backup_file, BACKUP_DIR)
-
 
     # copies from vps-setup/config/file to ~/.examplerc
     copy_file(file, dot_file)
   end
-end
-
-private
-
-# Defaults to Dir.home, can specify Dir, useful for backups
-def copy_file(file, new_file, dir = Dir.home)
-  new_file = File.join(dir, new_file)
-  FileUtils.cp(file, new_file)
-end
-
-def file_exists?(file, dir = Dir.home)
-  File.exist?(File.join(dir, file))
-end
-
-def create_backup_dir
-  backup = 'backup_dotfiles'
-
-  Dir.mkdir(File.join(Dir.home, backup)) unless Dir.exist?(backup)
-
-  backup_dir(backup)
-end
-
-def backup_dir(name)
-  Dir[File.join(Dir.home, name)]
 end
